Added project via upload.
Added the the full project. Is is a simple one the simulates a simple customer register system. Is uses .txt and .xml as source of DB infos for connection. The Template folder has the necessary scripts to be created in a SQL Server database for this project to work properly; it also has templates for the .txt and .xml files and an User Guide in brazilian portuguese.